requirements.txt
Django==5.2.6
djangorestframework==3.15.2
drf-spectacular==0.28.0
firebase-admin==6.5.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
Pillow==10.4.0
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
redis==5.2.0
celery==5.4.0
django-celery-beat==2.7.0
